Refactor CLI command code to a separate module
