--- src/config/block.rs
@@ -1,7 +1,0 @@
-use serde::Serialize;
-
-// Top level struct to hold the TOML data.
-#[derive(Serialize)]
-pub struct Config {
-    pub website_list_path: String,
-}
