+++ src/lib.rs
@@ -3,8 +3,14 @@ pub mod blocking {
     pub mod manage_websites;
 }
 pub mod config {
-    pub mod block;
+    
 }
+
 pub mod cli {
     pub mod cli_runner;
+    pub mod set_block_for_time_and_task;
+    pub mod commands {
+        pub mod reset;
+        pub mod setup;
+    }
 }
